refactor: delete store_anonymizer
