+++ great_expectations/core/usage_statistics/anonymizers/anonymizer.py
@@ -370,3 +370,19 @@ class Anonymizer:
         ] = anonymized_site_index_builder
 
         return anonymized_info_dict
+
+    def anonymize_store_info(self, store_name, store_obj):
+        anonymized_info_dict = {}
+        anonymized_info_dict["anonymized_name"] = self.anonymize(store_name)
+        store_backend_obj = store_obj.store_backend
+
+        self.anonymize_object_info(
+            object_=store_obj,
+            anonymized_info_dict=anonymized_info_dict,
+        )
+
+        anonymized_info_dict[
+            "anonymized_store_backend"
+        ] = self.anonymize_store_backend_info(store_backend_obj=store_backend_obj)
+
+        return anonymized_info_dict
